backend/requirements.txt
django
djangorestframework
pyrebase4
django-allauth
dj-rest-auth
djangorestframework-simplejwt
openai
django_filters
django-environ